requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
cffi==1.17.1
cryptography==44.0.1
Django==5.1.6
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
